Put value calculations into separate files.
